fix: 🐛 separate server and client components to resolve build errors
- Split WritingList into server component (data fetching) and WritingListView (client UI)
- Fix 'use client' directive issue causing Node.js module import errors
- Server components handle sharp/jsdom imports, client components handle UI state
